Updated Capture Blacklist disabled Ars Nouveau Drygmy and Jar blacklist files, in favor of the script. their entries appear to be duplicate to the EnderIO ones, so no merging has been done. changed local variables to constants and ensured they are prefixed as tags, for inheritance purposes
removed redundant entries from capture list, as some of them are covered by "forge:bosses"
added Sea Serpents to the capture list

in the future, the entity registry should be combed through. we may be able to recycle the generator script i made to handle MCE1's capture blacklist, or maybe it is not as necessary with the advanced features available.